log_event returns event timestamp

+++ Client/include/chronolog_client.h
@@ -105,7 +105,7 @@ class StoryHandle
 public:
     virtual  ~StoryHandle();
 
-    virtual int log_event(std::string const &) = 0;
+    virtual uint64_t log_event(std::string const &) = 0;
 
     virtual int playback_story(uint64_t start, uint64_t end, std::vector<Event> & playback_events) = 0;
 };
